Fixed a bug on the leader change protocol that would block all replicas if the delayed replica triggered again a leader change before the others finished the same protocol.
Identified a bug on the method "send" of class NettyClientServerCommunicationSystemServerSide, but its cause remains misterious: if invoked 2 times for the same message (why it sometimes is invoked 2 times for the same message is not known), the client may not receive a reply on its end, and if no sufficent replies are received, the client blocks. If locks are used in this method, the problem seems to disappear.

Removed various locks from the leader change protocol which were rendered useless in previous commits.

The Proposed thread now  unlocks by this order: I am the leader -> the consensus as finished  -> I have message to order