Fixed a bug that would manifest in a delayed replica: if a timeout occurred, the replica would be forever stopped, and the bug prevented it from triggering the state transfer protocol. With the bug gone, the replica will ask for the state.
Regarding the aforementioned bug, I added a parameter to the config file dubbed "timeoutHighmark". It should be lower than "paxosHighmark" and greater than "revivalHighmark". This will make the replica trigger the state transfer more quickly, given the conditions just described.

Fixed also a bug on the clients: when there were no "currentView" file, an exception would be thrown, and the clients would not read the host.config file. This no longer happens.

Finally, added some debugger messages in the leader change protocol and also in the servicereplica class.